Fetch a selected dataset from the DoReMiTra collection

This function fetches a SummarizedExperiment object from ExperimentHub corresponding to a dataset in the DoReMiTra package.

get_DoReMiTra_data(dataset_name, gene_symbol = FALSE)

Arguments

dataset_name

Character. The exact name of the dataset (e.g. "SE_Amundson_2008_ExVivo_GSE8917_GPL1708").

gene_symbol

Logical. default is FALSE. If TRUE, gene symbol will be assigned to rownames. If some gene symbols were found to be duplicated, gene symbol and the corresponding probe id will be appended together.

Value

A SummarizedExperiment object fetched from ExperimentHub.

Examples

get_DoReMiTra_data("SE_Amundson_2008_ExVivo_GSE8917_GPL1708")
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#> class: SummarizedExperiment 
#> dim: 43931 50 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(43931): BrightCorner NegativeControl ... BrightCorner
#>   BrightCorner
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(50): GSM225984 GSM225985 ... GSM226032 GSM226033
#> colData names(47): title geo_accession ... TaxonomyId Exp_setting