Fetch all datasets from the DoReMiTra collection

Retrieves all datasets available in the DoReMiTra package from ExperimentHub and returns them as a named list of SummarizedExperiment objects. This is useful for batch processing or exploring all curated radiation response datasets at once.

get_all_DoReMiTra_datasets(verbose = TRUE)

Arguments

verbose

Logical. Whether to print progress messages. Default is TRUE.

Value

A list storing all the SE objects

Examples

get_all_DoReMiTra_datasets()
#>  Loading dataset: SE_Amundson_2019_InVivo_GSE124612_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Aryankalayil_2018_InVivo_GSE104121_GPL10787
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Aryankalayil_2018_InVivo_GSE104121_GPL21163
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Park_2017_ExVivo_GSE102971_GPL10332_MacacaMulatta
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Park_2017_ExVivo_GSE102971_GPL10332_HomoSapiens
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Vasilyev_2017_ExVivo_GSE97000_GPL17077
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Paul_2013_ExVivo_GSE44201_GPL6480
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Paul_2013_ExVivo_GSE44201_GPL6848
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Nosel_2013_ExVivo_GSE43151_GPL13497
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Girardi_2012_ExVivo_GSE20173_GPL6480
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Amundson_2011_ExVivo_GSE23515_GPL6480
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Gruel_2008_ExVivo_GSE6978_GPL4803
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Rouchka_2019_ExVivo_GSE63952_GPL15207
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Amundson_2008_ExVivo_GSE8917_GPL1708
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Manikandan_2014_ExVivo_GSE36355_GPL6883
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Lee_2013_ExVivo_GSE44245_GPL570
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Rouchka_2015_ExVivo_GSE64375_GPL6244
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Ankermit_2015_ExVivo_GSE55953_GPL14550
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2018_InVivo_GSE113509_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2017_InVivo_GSE85323_GPL10333
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2017_ExVivo_GSE90909_GPL13497
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Paul_2010_InVivo_GSE23393_GPL6480
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Ghandhi_2015_ExVivo_GSE65292_GPL13497
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Flores_2009_ExVivo_GSE15341_GPL8332
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Amundson_2011_InVivo_GSE20162_GPL6480
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2023_InVivo_GSE196400_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2022_InVivo_GSE184361_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2021_InVivo_GSE132559_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Yamaguchi_2020_InVivo_GSE137192_GPL1261
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Mukherjee_2019_InVivo_GSE114142_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Amundson_2018_InVivo_GSE101402_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Amundson_2018_InVivo_GSE99176_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Paul_2015_InVivo_GSE62623_GPL10333
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Ghandhi_2018_InVivo_GSE84898_GPL13497
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Broustas_2021_InVivo_GSE133451_GPL11202
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#>  Loading dataset: SE_Salah_2025_ExVivo
#> see ?DoReMiTra and browseVignettes('DoReMiTra') for documentation
#> loading from cache
#> $SE_Amundson_2019_InVivo_GSE124612_GPL11202
#> class: SummarizedExperiment 
#> dim: 17920 248 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(17920): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG05_66023
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(248): GSM3537608 GSM3537609 ... GSM3537854 GSM3537855
#> colData names(51): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Aryankalayil_2018_InVivo_GSE104121_GPL10787
#> class: SummarizedExperiment 
#> dim: 30629 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(30629): A_30_P01017428 A_30_P01017437 ... A_66_P140954
#>   A_66_P140973
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(24): GSM2790225 GSM2790226 ... GSM2790247 GSM2790248
#> colData names(53): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Aryankalayil_2018_InVivo_GSE104121_GPL21163
#> class: SummarizedExperiment 
#> dim: 30629 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(30629): A_30_P01017428 A_30_P01017437 ... A_66_P140954
#>   A_66_P140973
#> rowData names(17): ID CONTROL_TYPE ... SPOT_ID SYMBOL
#> colnames(24): GSM2790249 GSM2790250 ... GSM2790271 GSM2790272
#> colData names(53): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Park_2017_ExVivo_GSE102971_GPL10332_MacacaMulatta
#> class: SummarizedExperiment 
#> dim: 43118 67 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(43118): A_23_P146146 A_23_P42935 ... A_23_P121875 A_23_P27810
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(67): GSM2751931 GSM2751932 ... GSM2751996 GSM2751997
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Park_2017_ExVivo_GSE102971_GPL10332_HomoSapiens
#> class: SummarizedExperiment 
#> dim: 43118 100 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(43118): A_23_P146146 A_23_P42935 ... A_23_P121875 A_23_P27810
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(100): GSM2751831 GSM2751832 ... GSM2751929 GSM2751930
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Vasilyev_2017_ExVivo_GSE97000_GPL17077
#> class: SummarizedExperiment 
#> dim: 36204 12 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(36204): A_19_P00315459 A_19_P00315492 ... A_33_P3897698
#>   A_33_P3897734
#> rowData names(17): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(12): GSM2549665 GSM2549666 ... GSM2549675 GSM2549676
#> colData names(55): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Paul_2013_ExVivo_GSE44201_GPL6480
#> class: SummarizedExperiment 
#> dim: 41094 25 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(41094): (+)E1A_r60_1 (+)E1A_r60_3 ... RC8 RC9
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(25): GSM1080614 GSM1080615 ... GSM1080662 GSM1080663
#> colData names(51): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Paul_2013_ExVivo_GSE44201_GPL6848
#> class: SummarizedExperiment 
#> dim: 41015 50 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(41015): (+)E1A_r60_1 (+)E1A_r60_3 ... A_32_P99933 A_32_P99942
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(50): GSM1080589 GSM1080590 ... GSM1080647 GSM1080648
#> colData names(52): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Nosel_2013_ExVivo_GSE43151_GPL13497
#> class: SummarizedExperiment 
#> dim: 19246 121 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(19246): A_23_P100001 A_23_P100074 ... A_33_P3888568
#>   A_33_P3893191
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(121): GSM1057397 GSM1057398 ... GSM1057516 GSM1057517
#> colData names(44): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Girardi_2012_ExVivo_GSE20173_GPL6480
#> class: SummarizedExperiment 
#> dim: 29706 20 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(29706): A_23_P100001 A_23_P100011 ... A_32_P99902 A_32_P99933
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(20): GSM506087 GSM506088 ... GSM506115 GSM506116
#> colData names(51): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Amundson_2011_ExVivo_GSE23515_GPL6480
#> class: SummarizedExperiment 
#> dim: 41093 95 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(41093): (+)E1A_r60_1 (+)E1A_r60_3 ... RC8 RC9
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(95): GSM575908 GSM575909 ... GSM576001 GSM576002
#> colData names(56): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Gruel_2008_ExVivo_GSE6978_GPL4803
#> class: SummarizedExperiment 
#> dim: 26496 114 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(26496): 1.1.1 1.1.10 ... 9.9.8 9.9.9
#> rowData names(9): ID Probe ... SPOT_ID SYMBOL
#> colnames(114): GSM160297 GSM160299 ... GSM160852 GSM160853
#> colData names(52): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Rouchka_2019_ExVivo_GSE63952_GPL15207
#> class: SummarizedExperiment 
#> dim: 49395 100 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(49395): 11715100_at 11715101_s_at ... AFFX-TrpnX-5_at
#>   AFFX-TrpnX-M_at
#> rowData names(37): ID GeneChip.Array ... SPOT_ID SYMBOL
#> colnames(100): GSM1561166 GSM1561167 ... GSM1561264 GSM1561265
#> colData names(52): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Amundson_2008_ExVivo_GSE8917_GPL1708
#> class: SummarizedExperiment 
#> dim: 43931 50 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(43931): BrightCorner NegativeControl ... BrightCorner
#>   BrightCorner
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(50): GSM225984 GSM225985 ... GSM226032 GSM226033
#> colData names(47): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Manikandan_2014_ExVivo_GSE36355_GPL6883
#> class: SummarizedExperiment 
#> dim: 24526 16 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(24526): ILMN_1343291 ILMN_1651209 ... ILMN_2415979
#>   ILMN_2416019
#> rowData names(28): ID Species ... GB_ACC SYMBOL
#> colnames(16): GSM889492 GSM889493 ... GSM889506 GSM889507
#> colData names(43): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Lee_2013_ExVivo_GSE44245_GPL570
#> class: SummarizedExperiment 
#> dim: 54675 6 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(54675): 1007_s_at 1053_at ... AFFX-TrpnX-5_at AFFX-TrpnX-M_at
#> rowData names(17): ID GB_ACC ... Gene.Ontology.Molecular.Function
#>   SYMBOL
#> colnames(6): GSM1081193 GSM1081194 ... GSM1081197 GSM1081198
#> colData names(42): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Rouchka_2015_ExVivo_GSE64375_GPL6244
#> class: SummarizedExperiment 
#> dim: 33297 16 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(33297): 7892501 7892502 ... 8180417 8180418
#> rowData names(13): ID GB_LIST ... category SYMBOL
#> colnames(16): GSM1569806 GSM1569807 ... GSM1569820 GSM1569821
#> colData names(51): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Ankermit_2015_ExVivo_GSE55953_GPL14550
#> class: SummarizedExperiment 
#> dim: 42545 28 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(42545): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG10_236652
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(28): GSM1349234 GSM1349235 ... GSM1349260 GSM1349261
#> colData names(47): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2018_InVivo_GSE113509_GPL11202
#> class: SummarizedExperiment 
#> dim: 14243 18 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(14243): (+)E1A_r60_a104 (+)E1A_r60_a107 ... ETG10_236652
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(18): GSM3108171 GSM3108172 ... GSM3108205 GSM3108206
#> colData names(44): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2017_InVivo_GSE85323_GPL10333
#> class: SummarizedExperiment 
#> dim: 44397 72 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(44397): GE_BrightCorner DarkCorner ... GE_BrightCorner
#>   GE_BrightCorner
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(72): GSM2264695 GSM2264696 ... GSM2264765 GSM2264766
#> colData names(43): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2017_ExVivo_GSE90909_GPL13497
#> class: SummarizedExperiment 
#> dim: 12289 92 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(12289): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG10_234183
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(92): GSM2417353 GSM2417354 ... GSM2417443 GSM2417444
#> colData names(44): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Paul_2010_InVivo_GSE23393_GPL6480
#> class: SummarizedExperiment 
#> dim: 41093 16 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(41093): (+)E1A_r60_1 (+)E1A_r60_3 ... RC8 RC9
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(16): GSM573686 GSM573687 ... GSM573700 GSM573701
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Ghandhi_2015_ExVivo_GSE65292_GPL13497
#> class: SummarizedExperiment 
#> dim: 12073 35 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(12073): (+)E1A_r60_3 (+)E1A_r60_a104 ... ETG07_105829
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(35): GSM1591727 GSM1591728 ... GSM1591760 GSM1591761
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Flores_2009_ExVivo_GSE15341_GPL8332
#> class: SummarizedExperiment 
#> dim: 12604 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(12604): Control_Cot-1 Control_pdA ... SOS_04_24uM SOS_04_40uM
#> rowData names(14): ID Gene.Symbol ... SPOT_ID SYMBOL
#> colnames(24): GSM385284 GSM385285 ... GSM385306 GSM385307
#> colData names(55): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Amundson_2011_InVivo_GSE20162_GPL6480
#> class: SummarizedExperiment 
#> dim: 41093 66 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(41093): (+)E1A_r60_1 (+)E1A_r60_3 ... RC8 RC9
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(66): GSM505930 GSM505931 ... GSM505994 GSM505995
#> colData names(52): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2023_InVivo_GSE196400_GPL11202
#> class: SummarizedExperiment 
#> dim: 12900 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(12900): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG08_142674
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(24): GSM5880264 GSM5880265 ... GSM5880286 GSM5880287
#> colData names(47): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2022_InVivo_GSE184361_GPL11202
#> class: SummarizedExperiment 
#> dim: 18417 30 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(18417): (+)E1A_r60_3 (+)E1A_r60_a104 ... ETG10_236652
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(30): GSM5585072 GSM5585073 ... GSM5585100 GSM5585101
#> colData names(45): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2021_InVivo_GSE132559_GPL11202
#> class: SummarizedExperiment 
#> dim: 17190 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(17190): (+)E1A_r60_3 (+)E1A_r60_a107 ... ETG07_105829
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(24): GSM3874658 GSM3874659 ... GSM3874680 GSM3874681
#> colData names(43): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Yamaguchi_2020_InVivo_GSE137192_GPL1261
#> class: SummarizedExperiment 
#> dim: 45101 17 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(45101): 1415670_at 1415671_at ... AFFX-TrpnX-5_at
#>   AFFX-TrpnX-M_at
#> rowData names(17): ID GB_ACC ... Gene.Ontology.Molecular.Function
#>   SYMBOL
#> colnames(17): GSM4072759 GSM4072760 ... GSM4072774 GSM4072775
#> colData names(50): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Mukherjee_2019_InVivo_GSE114142_GPL11202
#> class: SummarizedExperiment 
#> dim: 16909 20 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(16909): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG10_236652
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(20): GSM3135538 GSM3135539 ... GSM3135556 GSM3135557
#> colData names(45): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Amundson_2018_InVivo_GSE101402_GPL11202
#> class: SummarizedExperiment 
#> dim: 10214 24 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(10214): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG05_66023
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(24): GSM2702018 GSM2702019 ... GSM2702040 GSM2702041
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Amundson_2018_InVivo_GSE99176_GPL11202
#> class: SummarizedExperiment 
#> dim: 20251 32 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(20251): (+)E1A_r60_1 (+)E1A_r60_3 ... ETG10_234183
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(32): GSM2635156 GSM2635157 ... GSM2635186 GSM2635187
#> colData names(51): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Paul_2015_InVivo_GSE62623_GPL10333
#> class: SummarizedExperiment 
#> dim: 44397 48 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(44397): GE_BrightCorner DarkCorner ... GE_BrightCorner
#>   GE_BrightCorner
#> rowData names(21): ID COL ... SEQUENCE SYMBOL
#> colnames(48): GSM1530171 GSM1530172 ... GSM1530217 GSM1530218
#> colData names(50): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Ghandhi_2018_InVivo_GSE84898_GPL13497
#> class: SummarizedExperiment 
#> dim: 5885 20 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(5885): A_23_P100001 A_23_P100056 ... A_33_P3897698
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(20): GSM2253686 GSM2253687 ... GSM2253704 GSM2253705
#> colData names(45): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Broustas_2021_InVivo_GSE133451_GPL11202
#> class: SummarizedExperiment 
#> dim: 14471 20 
#> metadata(1): DoReMiTra
#> assays(1): exprs
#> rownames(14471): (+)E1A_r60_a107 (+)E1A_r60_a135 ... ETG10_13482
#>   GE_BrightCorner
#> rowData names(18): ID SPOT_ID ... SEQUENCE SYMBOL
#> colnames(20): GSM3908740 GSM3908741 ... GSM3908758 GSM3908759
#> colData names(49): title geo_accession ... TaxonomyId Exp_setting
#> 
#> $SE_Salah_2025_ExVivo
#> class: SummarizedExperiment 
#> dim: 62646 30 
#> metadata(8): tximetaInfo quantInfo ... assignRanges DoReMiTra
#> assays(3): counts abundance length
#> rownames(62646): ENSG00000000003 ENSG00000000005 ... ENSG00000293599
#>   ENSG00000293600
#> rowData names(3): gene_id tx_ids SYMBOL
#> colnames(30): 0Gy-2hrs-do1 1Gy-2hrs-do1 ... 2Gy-2hrs-do3 4Gy-6hrs-do3
#> colData names(10): Dose Time_point ... Exp_setting Platform
#>