List all available DoReMiTra datasets with associated key metadata information

Returns a metadata dataframe of all datasets available in the DoReMiTra collection, including details such as title, organism, radiation type, experimental setting, and accession numbers. Can optionally display extended metadata fields.

list_DoReMiTra_datasets(show_all_fields = FALSE)

Arguments

show_all_fields

Logical. If TRUE, it returns all the metadata information

Value

A data.frame with metadata for each dataset

Examples

list_DoReMiTra_datasets()
#>                                                 Dataset  RadiationType
#> 1            SE_Amundson_2019_InVivo_GSE124612_GPL11202      gamma ray
#> 2        SE_Aryankalayil_2018_InVivo_GSE104121_GPL10787      gamma ray
#> 3        SE_Aryankalayil_2018_InVivo_GSE104121_GPL21163      gamma ray
#> 4  SE_Park_2017_ExVivo_GSE102971_GPL10332_MacacaMulatta      gamma ray
#> 5    SE_Park_2017_ExVivo_GSE102971_GPL10332_HomoSapiens      gamma ray
#> 6             SE_Vasilyev_2017_ExVivo_GSE97000_GPL17077      gamma ray
#> 7                  SE_Paul_2013_ExVivo_GSE44201_GPL6480      gamma ray
#> 8                  SE_Paul_2013_ExVivo_GSE44201_GPL6848      gamma ray
#> 9                SE_Nosel_2013_ExVivo_GSE43151_GPL13497      gamma ray
#> 10              SE_Girardi_2012_ExVivo_GSE20173_GPL6480      gamma ray
#> 11             SE_Amundson_2011_ExVivo_GSE23515_GPL6480      gamma ray
#> 12                 SE_Gruel_2008_ExVivo_GSE6978_GPL4803      gamma ray
#> 13             SE_Rouchka_2019_ExVivo_GSE63952_GPL15207      gamma ray
#> 14              SE_Amundson_2008_ExVivo_GSE8917_GPL1708      gamma ray
#> 15           SE_Manikandan_2014_ExVivo_GSE36355_GPL6883      gamma ray
#> 16                   SE_Lee_2013_ExVivo_GSE44245_GPL570      gamma ray
#> 17              SE_Rouchka_2015_ExVivo_GSE64375_GPL6244      gamma ray
#> 18            SE_Ankermit_2015_ExVivo_GSE55953_GPL14550      gamma ray
#> 19           SE_Broustas_2018_InVivo_GSE113509_GPL11202 Neutron; X-ray
#> 20            SE_Broustas_2017_InVivo_GSE85323_GPL10333 Neutron; X-ray
#> 21            SE_Broustas_2017_ExVivo_GSE90909_GPL13497 Neutron; X-ray
#> 22                 SE_Paul_2010_InVivo_GSE23393_GPL6480          X-ray
#> 23             SE_Ghandhi_2015_ExVivo_GSE65292_GPL13497          X-ray
#> 24               SE_Flores_2009_ExVivo_GSE15341_GPL8332          X-ray
#> 25             SE_Amundson_2011_InVivo_GSE20162_GPL6480          X-ray
#> 26           SE_Broustas_2023_InVivo_GSE196400_GPL11202          X-ray
#> 27           SE_Broustas_2022_InVivo_GSE184361_GPL11202          X-ray
#> 28           SE_Broustas_2021_InVivo_GSE132559_GPL11202          X-ray
#> 29           SE_Yamaguchi_2020_InVivo_GSE137192_GPL1261          X-ray
#> 30          SE_Mukherjee_2019_InVivo_GSE114142_GPL11202          X-ray
#> 31           SE_Amundson_2018_InVivo_GSE101402_GPL11202          X-ray
#> 32            SE_Amundson_2018_InVivo_GSE99176_GPL11202          X-ray
#> 33                SE_Paul_2015_InVivo_GSE62623_GPL10333          X-ray
#> 34             SE_Ghandhi_2018_InVivo_GSE84898_GPL13497          X-ray
#> 35           SE_Broustas_2021_InVivo_GSE133451_GPL11202          X-ray
#> 36                                 SE_Salah_2025_ExVivo          X-ray
#>          Organism ExpSetting Accession
#> 1    Mus musculus     InVivo GSE124612
#> 2    Mus musculus     InVivo GSE104121
#> 3    Mus musculus     InVivo GSE104121
#> 4  Macaca mulatta     ExVivo GSE102971
#> 5    Homo sapiens     ExVivo GSE102971
#> 6    Homo sapiens     ExVivo  GSE97000
#> 7    Homo sapiens     ExVivo  GSE44201
#> 8    Homo sapiens     ExVivo  GSE44201
#> 9    Homo sapiens     ExVivo  GSE43151
#> 10   Homo sapiens     ExVivo  GSE20173
#> 11   Homo sapiens     ExVivo  GSE23515
#> 12   Homo sapiens     ExVivo   GSE6978
#> 13   Homo sapiens     ExVivo  GSE63952
#> 14   Homo sapiens     ExVivo   GSE8917
#> 15   Homo sapiens     ExVivo  GSE36355
#> 16   Homo sapiens     ExVivo  GSE44245
#> 17   Homo sapiens     ExVivo  GSE64375
#> 18   Homo sapiens     ExVivo  GSE55953
#> 19   Mus musculus     InVivo GSE113509
#> 20   Mus musculus     InVivo  GSE85323
#> 21   Homo sapiens     ExVivo  GSE90909
#> 22   Homo sapiens     InVivo  GSE23393
#> 23   Homo sapiens     ExVivo  GSE65292
#> 24   Homo sapiens     ExVivo  GSE15341
#> 25   Homo sapiens     InVivo  GSE20162
#> 26   Mus musculus     InVivo GSE196400
#> 27   Mus musculus     InVivo GSE184361
#> 28   Mus musculus     InVivo GSE132559
#> 29   Mus musculus     InVivo GSE137192
#> 30   Mus musculus     InVivo GSE114142
#> 31   Mus musculus     InVivo GSE101402
#> 32   Mus musculus     InVivo  GSE99176
#> 33   Mus musculus     InVivo  GSE62623
#> 34 Macaca mulatta     InVivo  GSE84898
#> 35   Mus musculus     InVivo GSE133451
#> 36   Homo sapiens     ExVivo      <NA>