Search DoReMiTra datasets by metadata filters

Filters the available DoReMiTra datasets using metadata fields such as radiation type, organism, or experimental setting. This function helps narrow down datasets of interest before fetching them.

search_DoReMiTra_datasets(
  radiation_type = NULL,
  organism = NULL,
  exp_setting = NULL
)

Arguments

radiation_type

Character string (optional). Filter datasets by radiation type (e.g., "x-ray", "neutron").

organism

Character string (optional). Filter by organism (e.g., "Homo sapiens").

exp_setting

Character string (optional). Filter by experimental setting (e.g., "in vivo", "ex vivo").

Value

a vector with the names of the se objects matching the inclusion criteria. If none match, returns an empty vector with a message.

Examples

search_DoReMiTra_datasets()
#> Radiation type not provided - searching across all radiation types.
#> Organism not provided - searching across all species.
#> Experimental setting not provided - searching across all settings (in vivo, ex vivo, etc.).
#> 
#> Matching datasets found: 36
#> To retrieve one or more of these datasets, you can use:
#> se_name1 <- get_DoReMiTra_data("SE_Amundson_2019_InVivo_GSE124612_GPL11202")
#> se_name2 <- get_DoReMiTra_data("SE_Aryankalayil_2018_InVivo_GSE104121_GPL10787")
#> 
#> For more details, please refer to `?get_DoReMiTra_data`
#>  [1] "SE_Amundson_2019_InVivo_GSE124612_GPL11202"          
#>  [2] "SE_Aryankalayil_2018_InVivo_GSE104121_GPL10787"      
#>  [3] "SE_Aryankalayil_2018_InVivo_GSE104121_GPL21163"      
#>  [4] "SE_Park_2017_ExVivo_GSE102971_GPL10332_MacacaMulatta"
#>  [5] "SE_Park_2017_ExVivo_GSE102971_GPL10332_HomoSapiens"  
#>  [6] "SE_Vasilyev_2017_ExVivo_GSE97000_GPL17077"           
#>  [7] "SE_Paul_2013_ExVivo_GSE44201_GPL6480"                
#>  [8] "SE_Paul_2013_ExVivo_GSE44201_GPL6848"                
#>  [9] "SE_Nosel_2013_ExVivo_GSE43151_GPL13497"              
#> [10] "SE_Girardi_2012_ExVivo_GSE20173_GPL6480"             
#> [11] "SE_Amundson_2011_ExVivo_GSE23515_GPL6480"            
#> [12] "SE_Gruel_2008_ExVivo_GSE6978_GPL4803"                
#> [13] "SE_Rouchka_2019_ExVivo_GSE63952_GPL15207"            
#> [14] "SE_Amundson_2008_ExVivo_GSE8917_GPL1708"             
#> [15] "SE_Manikandan_2014_ExVivo_GSE36355_GPL6883"          
#> [16] "SE_Lee_2013_ExVivo_GSE44245_GPL570"                  
#> [17] "SE_Rouchka_2015_ExVivo_GSE64375_GPL6244"             
#> [18] "SE_Ankermit_2015_ExVivo_GSE55953_GPL14550"           
#> [19] "SE_Broustas_2018_InVivo_GSE113509_GPL11202"          
#> [20] "SE_Broustas_2017_InVivo_GSE85323_GPL10333"           
#> [21] "SE_Broustas_2017_ExVivo_GSE90909_GPL13497"           
#> [22] "SE_Paul_2010_InVivo_GSE23393_GPL6480"                
#> [23] "SE_Ghandhi_2015_ExVivo_GSE65292_GPL13497"            
#> [24] "SE_Flores_2009_ExVivo_GSE15341_GPL8332"              
#> [25] "SE_Amundson_2011_InVivo_GSE20162_GPL6480"            
#> [26] "SE_Broustas_2023_InVivo_GSE196400_GPL11202"          
#> [27] "SE_Broustas_2022_InVivo_GSE184361_GPL11202"          
#> [28] "SE_Broustas_2021_InVivo_GSE132559_GPL11202"          
#> [29] "SE_Yamaguchi_2020_InVivo_GSE137192_GPL1261"          
#> [30] "SE_Mukherjee_2019_InVivo_GSE114142_GPL11202"         
#> [31] "SE_Amundson_2018_InVivo_GSE101402_GPL11202"          
#> [32] "SE_Amundson_2018_InVivo_GSE99176_GPL11202"           
#> [33] "SE_Paul_2015_InVivo_GSE62623_GPL10333"               
#> [34] "SE_Ghandhi_2018_InVivo_GSE84898_GPL13497"            
#> [35] "SE_Broustas_2021_InVivo_GSE133451_GPL11202"          
#> [36] "SE_Salah_2025_ExVivo"